+++ crates/tzlint_rules/src/rules/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod ja_no_mixed_period;
 pub mod ja_no_redundant_expression;
+pub mod ja_prh;
 pub mod max_kanji_continuous_len;
 pub mod max_ten;
 pub mod no_double_negative_ja;